Modify README to use common instructions (#46)
Docker Compose support used in this guide is common to multiple
Spring Guides. The description of Docker Compose Support in
Spring Boot should be a common file that is referenced with a
link instead of using the same text in multiple guides.

Linking to a shared file will make maintaining multiple guides
easier and faster if the text can be modified in a single
location.